Toan: Done pay-salary

+++ EPayroll_BE/ViewModels/PayTypeAmountViewModel.cs
@@ -9,7 +9,7 @@ namespace EPayroll_BE.ViewModels
     public class PayTypeAmountViewModel
     {
         public Guid Id { get; set; }
-        public float Amount { get; set; }
+        public long Amount { get; set; }
 
 
         public SalaryLevelViewModel SalaryLevel { get; set; }
@@ -20,7 +20,7 @@ namespace EPayroll_BE.ViewModels
     public class PayTypeAmountCreateModel
     {
         [Required]
-        public float Amount { get; set; }
+        public long Amount { get; set; }
 
         public Guid SalaryLevelId { get; set; }
         public Guid PayTypeId { get; set; }
